+++ packages/berlin/src/components/comment-card/CommentCard.tsx
@@ -17,7 +17,7 @@ import { Body } from '../typography/Body.styled';
 import { FlexRow } from '../containers/FlexRow.styled';
 import Button from '../button';
 import Dialog from '../dialog';
-import IconButton from '../iconButton';
+import IconButton from '../icon-button';
 
 // Styled Components
 import { Card, FormattedDate, Username } from './CommentCard.styled';
