+++ packages/berlin/src/components/option-card/OptionCard.styled.tsx
