+++ packages/berlin/src/components/result-card/ResultCard.styled.tsx
