+++ packages/ui/README.md
@@ -1,3 +1,6 @@
 # UI package
 
 This is the UI package for pluraltools.
+
+folders on kebab-case
+components on PascalCase
